fixed wordpress connection to use redirects

+++ app/Http/Controllers/Connections/WordpressController.php
@@ -12,59 +12,58 @@ class WordpressController extends BaseConnectionController
 {
     public function __construct()
     {
-        $this->auth = new WordPressAuth;
-        $this->api = new WordPressAPI;
+        $this->auth = new WordPressAuth();
+        $this->api = new WordPressAPI();
     }
 
     public function callback(Request $request)
     {
         $code = $request->input('code');
         $token = $this->auth->codeForToken($code);
+        $redirectRoute = $this->redirectRoute();
 
         if (collect($token)->has('error')) {
             $connection = $this->getSessionConnection();
-            $connection->delete();
+
+            if ($connection) {
+                $connection->delete();
+            }
 
             $this->cleanSessionConnection();
 
-            return redirect()->route('connectionIndex')->with([
-                'flash_message' => "There was an error with your authentication, please try again",
+            return redirect()->route($redirectRoute)->with([
+                'flash_message' => 'There was an error with your authentication, please try again',
                 'flash_message_type' => 'danger',
-                'flash_message_important' => true
+                'flash_message_important' => true,
             ]);
         }
 
-        $connection = $this->saveConnectionSettings($token);
+        $tokenArray = (array) $token;
+        $connection = $this->saveConnection($tokenArray);
 
-        return redirect()->route('connectionIndex')->with([
-            'flash_message' => "Wordpress connection <strong>" . $connection->name . "</strong> created successfully.",
+        return redirect()->route($redirectRoute)->with([
+            'flash_message' => 'Wordpress connection <strong>'.$connection->name.'</strong> created successfully.',
             'flash_message_type' => 'success',
-            'flash_message_important' => true
+            'flash_message_important' => true,
         ]);
     }
 
-    private function saveConnectionSettings($token)
+    protected function saveConnection(array $tokenArray, $providerSlug = null)
     {
-        $connectionData = Session::get('connection_data');
-        $metaData = $connectionData['meta_data'];
+        $metaData = $this->getSessionConnectionMetadata();
 
         $url = $this->formatUrl($metaData['url']);
-
         $blogInfo = $this->api->blogInfo($url);
 
-        $token->blog_url = $url;
-        $token->blog_id = $blogInfo->ID;
+        $tokenArray['blog_url'] = $url;
+        $tokenArray['blog_id'] = $blogInfo->ID;
 
-        $connection = $this->getSessionConnection();
-        $tokenArray = (array) $token;
-        $connection->saveSettings($tokenArray);
-
-        return $connection;
+        return parent::saveConnection($tokenArray, 'wordpress');
     }
 
     private function formatUrl($url)
     {
-        $disallowed = [ 'http://', 'https://' ];
+        $disallowed = ['http://', 'https://'];
 
         foreach ($disallowed as $d) {
             if (strpos($url, $d) === 0) {
